requirements.txt
Django
djangorestframework
python-dotenv
requests